requirements.txt
torch==1.13.0
tqdm==4.64.0
scikit_image==0.18.1
easydict==1.9
open3d==0.17.0
numpy==1.17.4
Click==7.0
natsort==7.1.1
matplotlib==3.1.2
pytorch_lightning==1.8.4.post0
PyYAML==6.0.1
scipy==1.11.1
skimage==0.0
